improved logging in bash scripts

+++ vm_scripts/evaluate_models.sh
@@ -2,9 +2,9 @@
 
 CONTAINER_NAME=$1
 QUEUE_NAME=$2
-MODEL_DIR=/home/trojai/data/
+MODEL_DIR=/home/trojai/data
 
-ACTIVE_DIR=/home/trojai/active/
+ACTIVE_DIR=/home/trojai/active
 
 CONTAINER_EXEC=/mnt/scratch/$CONTAINER_NAME
 RESULT_DIR=/mnt/scratch/results
@@ -41,8 +41,10 @@ do
 				singularity run --contain -B /mnt/scratch -B $ACTIVE_DIR --nv $CONTAINER_EXEC --model_filepath $ACTIVE_DIR/model.pt --result_filepath $ACTIVE_DIR/result.txt --scratch_dirpath $SCRATCH_DIR --examples_dirpath $ACTIVE_DIR/example_data >> $RESULT_DIR/$CONTAINER_NAME.out 2>&1
 				echo "Finished executing, returned status code: $?" >> $RESULT_DIR/$CONTAINER_NAME.out 2>&1
 
-				# copy result back to real output filename based on model name
-				cp $ACTIVE_DIR/result.txt  $RESULT_DIR/$MODEL.txt
+				if [[ -f $ACTIVE_DIR/result.txt ]]; then
+					# copy result back to real output filename based on model name
+					cp $ACTIVE_DIR/result.txt  $RESULT_DIR/$MODEL.txt
+				fi
 			fi
 		fi
 	fi
